Paranoid'ify the manifest
This could be an outcry for developers, but that's not gonna help to
be honest, because those don't attract bright minds. Bright minds are
attracted to innovation, so we are gonna try do that once more. So,
here's to fresh beginnings and not letting anyone tell us we are dead.

Change-Id: Idc4c65ac8d0fa77a43aa586e75b8086c600b8b46